Add highlighting rows on hover and update tests snapshots
